FIX (payments): Harden provider identity, webhooks, and lifecycle guards
- Fail closed on unsafe production Google webhook verification settings and require explicit Pub/Sub audience validation
- Preserve provider identity boundaries by keeping Google purchase tokens in token fields and avoiding token-as-transaction-id fallback paths
- Strengthen Google lifecycle processing to match subscriptions by purchase token and avoid cross-user or same-SKU state bleed
- Add claim-token fencing, processed-delivery guards, stale-event suppression, and terminal-state protections across webhook and scheduler flows
- Scope reconciliation, pause scheduling, orphan cleanup, and payment/subscription lookups through app-aware query paths for RLS compatibility
- Keep Google one-time purchase ACK identity rows even when provider price data is missing, while storing unknown money as NULL instead of fake values
- Reject subscription resume requests unless the local status is paused or cancelled before any provider-side resume call
- Redact sensitive database/provider errors and reduce PII exposure in logs, callbacks, and diagnostics
- Add migrations for bigint payment/price fields and nullable payment amounts required by ACK identity rows
- Add regression coverage for Google token isolation, webhook claim fencing, signature bypass prevention, invalid resume ordering, and OTP missing-price ACK persistence
- Update GPBI OTP suite wiring, cleanup, README coverage, release notes, and design docs for the hardened payment behavior

+++ src/application/subscription_actions_types.rs
@@ -50,7 +50,7 @@ pub struct PriceStepUpRequest {
 #[derive(Debug, Serialize)]
 pub struct PriceStepUpAcceptResponse {
     pub accepted: bool,
-    pub new_price_cents: i32,
+    pub new_price_cents: i64,
 }
 
 #[derive(Debug, Serialize)]
