+++ src/ports/types.rs
@@ -56,7 +56,7 @@ pub enum SubscriptionWebhookTransition {
     PaymentFailed,
     PendingPurchaseCancelled,
     PriceStepUp {
-        google_new_price_cents: Option<i32>,
+        google_new_price_cents: Option<i64>,
         google_price_step_up_consent_deadline: Option<chrono::DateTime<chrono::Utc>>,
     },
     PendingPriceChange {
@@ -84,7 +84,7 @@ pub struct WebhookPaymentRecordRequest<'a> {
     pub ack_required: bool,
     pub subscription_id: Option<&'a str>,
     pub product_id: Option<&'a str>,
-    pub amount_cents: i32,
+    pub amount_cents: i64,
     pub currency: Option<&'a str>,
     pub status: &'a str,
 }
@@ -155,7 +155,7 @@ pub struct OwnedWebhookPaymentRecord {
     pub ack_required: bool,
     pub subscription_id: Option<String>,
     pub product_id: Option<String>,
-    pub amount_cents: i32,
+    pub amount_cents: i64,
     pub currency: Option<String>,
     pub status: String,
 }
